+++ nurseryRhyme/safePlace.c
@@ -2,7 +2,7 @@
 #include "errorCode.h"
 #include "list.h"
 
-int findSafePlace(const int mortality, const int numberOfVictims, int *errorCode) {
+size_t findSafePlace(const size_t mortality, const size_t numberOfVictims, int *errorCode) {
     if (mortality < 1 || numberOfVictims < 1) {
         *errorCode = INCORRECT_ARGUMENTS_PASSED_TO_FUNCTION;
         return -1;
@@ -11,7 +11,7 @@ int findSafePlace(const int mortality, const int numberOfVictims, int *errorCode
     if (*errorCode != NO_ERRORS) {
         return -1;
     }
-    for (int i = 0; i < numberOfVictims; ++i) {
+    for (size_t i = 0; i < numberOfVictims; ++i) {
         addToList(list, i, errorCode);
         if (*errorCode != NO_ERRORS) {
             deleteList(&list, errorCode);
@@ -19,7 +19,7 @@ int findSafePlace(const int mortality, const int numberOfVictims, int *errorCode
         }
     }
     Node *node = getHead(list, errorCode);
-    for (int i = 0, k = 0; i < numberOfVictims - 1; ++k) {
+    for (size_t i = 0, k = 0; i < numberOfVictims - 1; ++k) {
         if (k == mortality - 1) {
             deleteNext(list, node, errorCode);
             ++i;
