+++ nurseryRhyme/safePlace.h
@@ -1,4 +1,6 @@
 #pragma once
 
+#include <stdlib.h>
+
 // finds the safe place in the counting game
-int findSafePlace(int mortality, int numberOfVictims, int *errorCode);
+size_t findSafePlace(size_t mortality, size_t numberOfVictims, int *errorCode);
